# routes/api.php
<?php

use App\Http\Controllers\BrandDeviceController;
use App\Http\Controllers\BrowserController;
use App\Http\Controllers\AuditEventController;
use App\Http\Controllers\ArtifactDescriptorController;
use App\Http\Controllers\AgentRegistrationController;
use App\Http\Controllers\AssetImpactController;
use App\Http\Controllers\AssetVersionController;
use App\Http\Controllers\CapabilityController;
use App\Http\Controllers\CostumerController;
use App\Http\Controllers\EnvironmentController;
use App\Http\Controllers\HeaderController;
use App\Http\Controllers\IdeliumClController;
use App\Http\Controllers\IdeliumInsertClController;
use App\Http\Controllers\ImportSeleniumController;
use App\Http\Controllers\LocationController;
use App\Http\Controllers\LoginController;
use App\Http\Controllers\ModelDeviceController;
use App\Http\Controllers\OsController;
use App\Http\Controllers\OidcWorkloadIdentityController;
use App\Http\Controllers\ParallelRunScheduleController;
use App\Http\Controllers\PerformedStepController;
use App\Http\Controllers\PerformedTestController;
use App\Http\Controllers\PerformedTestCycleController;
use App\Http\Controllers\PlatformController;
use App\Http\Controllers\PluginController;
use App\Http\Controllers\ProjectController;
use App\Http\Controllers\RoleController;
use App\Http\Controllers\ServiceAccountController;
use App\Http\Controllers\SideBarController;
use App\Http\Controllers\StatusController;
use App\Http\Controllers\StepController;
use App\Http\Controllers\TestController;
use App\Http\Controllers\TestCycleController;
use App\Http\Controllers\TestLauncherController;
use App\Http\Controllers\TypeController;
use App\Http\Controllers\UserController;
use App\Http\Controllers\VersionBrowserController;
use App\Http\Controllers\VersionOsController;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Route;
use Laravel\Sanctum\Http\Controllers\CsrfCookieController;

Route::get('sanctum/csrf-cookie', [CsrfCookieController::class, 'show'])
    ->name('csrf.show');
Route::post('login', [LoginController::class, 'login'])
    ->name('login');
Route::post('oidc/token-exchange', [OidcWorkloadIdentityController::class, 'exchange'])
    ->name('oidc.token-exchange');

Route::middleware(['auth:sanctum', 'tenant.context'])->group(function () {
    Route::get('me/capabilities', [CapabilityController::class, 'me'])
        ->name('capabilities.me');
    Route::post('logout', [LoginController::class, 'logout'])
        ->name('logout');
    /* menu */
    Route::get('menu/sidebar', [SideBarController::class, 'index'])
        ->name('sidebar.index');
    Route::get('menu/header', [HeaderController::class, 'index'])
        ->name('header.index');
    Route::put('menu/header/{idCostumer}', [HeaderController::class, 'changeCostumer'])
        ->name('header.changeCostumer');
    /* audit */
    Route::get('audit-events', [AuditEventController::class, 'index'])
        ->name('audit-events.index');
    /* roles */
    Route::get('admin/roles', [RoleController::class, 'index'])
        ->name('roles.index');
    /* profile */
    Route::get('admin/profile', [UserController::class, 'getuser'])
        ->name('accounts.getuser');
    Route::put('admin/profile', [UserController::class, 'updatePasswordUser'])
        ->name('accounts.updatePasswordUser');
    /* accounts */
    Route::get('admin/accounts', [UserController::class, 'index'])
        ->name('accounts.index');
    Route::post('admin/accounts', [UserController::class, 'store'])
        ->name('accounts.store');
    Route::put('admin/accounts/{idUser}', [UserController::class, 'update'])
        ->name('accounts.update');
    Route::delete('admin/accounts/{idUser}', [UserController::class, 'destroy'])
        ->name('accounts.destroy');
    /* costumers */
    Route::get('admin/costumers', [CostumerController::class, 'index'])
        ->name('costumers.index');
    Route::post('admin/costumers', [CostumerController::class, 'store'])
        ->name('costumers.store');
    Route::put('admin/costumers/{idCostumer}', [CostumerController::class, 'update'])
        ->name('costumers.update');
    Route::delete('admin/costumers/{idCostumer}', [CostumerController::class, 'destroy'])
        ->name('costumers.destroy');
    /* apikey */
    Route::get('admin/apikey', [CostumerController::class, 'getKey'])
        ->name('costumers.getKey');
    Route::put('admin/apikey', [CostumerController::class, 'updateKey'])
        ->name('costumers.updateKey');
    /* service accounts */
    Route::get('admin/service-accounts', [ServiceAccountController::class, 'index'])
        ->name('service-accounts.index');
    Route::post('admin/service-accounts', [ServiceAccountController::class, 'store'])
        ->name('service-accounts.store');
    Route::post('admin/service-accounts/{serviceAccount}/revoke', [ServiceAccountController::class, 'revoke'])
        ->name('service-accounts.revoke');
    /* agents */
    Route::get('admin/agents', [AgentRegistrationController::class, 'index'])
        ->name('agents.index');
    Route::put('admin/agents/{agentRegistration}/status', [AgentRegistrationController::class, 'updateStatus'])
        ->name('agents.updateStatus');
    /* projects */
    Route::resource('admin/projects', ProjectController::class);
    /* asset impact */
    Route::get(
        'admin/projects/{idProject}/asset-impact/{assetType}/{assetId}',
        [AssetImpactController::class, 'show']
    )->whereNumber('assetId')->name('assetimpact.show');
    /* asset versions */
    Route::get(
        'admin/projects/{idProject}/asset-versions/{assetType}/{assetId}',
        [AssetVersionController::class, 'index']
    )->whereNumber('assetId')->name('assetversions.index');
    Route::get(
        'admin/projects/{idProject}/asset-versions/{fromVersion}/diff/{toVersion}',
        [AssetVersionController::class, 'diff']
    )->whereNumber(['fromVersion', 'toVersion'])->name('assetversions.diff');
    Route::post(
        'admin/projects/{idProject}/asset-versions/{assetVersion}/review-events',
        [AssetVersionController::class, 'transitionReview']
    )->whereNumber('assetVersion')->name('assetversions.transitionReview');
    Route::get(
        'admin/projects/{idProject}/asset-versions/{assetVersion}',
        [AssetVersionController::class, 'show']
    )->whereNumber('assetVersion')->name('assetversions.show');
    /* parallel runs */
    Route::get('admin/projects/{idProject}/parallel-runs', [ParallelRunScheduleController::class, 'index'])
        ->name('parallelruns.index');
    Route::post('admin/projects/{idProject}/parallel-runs/matrix', [ParallelRunScheduleController::class, 'storeMatrix'])
        ->name('parallelruns.storeMatrix');
    Route::post('admin/projects/{idProject}/parallel-runs', [ParallelRunScheduleController::class, 'store'])
        ->name('parallelruns.store');
    Route::get('admin/projects/{idProject}/parallel-runs/{parallelRun}', [ParallelRunScheduleController::class, 'show'])
        ->name('parallelruns.show');
    Route::post('admin/projects/{idProject}/parallel-runs/{parallelRun}/claim', [ParallelRunScheduleController::class, 'claimWorker'])
        ->name('parallelruns.claimWorker');
    Route::post('admin/projects/{idProject}/parallel-runs/{parallelRun}/workers/{workerId}/heartbeat', [ParallelRunScheduleController::class, 'heartbeatWorker'])
        ->name('parallelruns.heartbeatWorker');
    Route::put('admin/projects/{idProject}/parallel-runs/{parallelRun}/workers/{workerId}', [ParallelRunScheduleController::class, 'updateWorker'])
        ->name('parallelruns.updateWorker');
    Route::post('admin/projects/{idProject}/parallel-runs/{parallelRun}/cancel', [ParallelRunScheduleController::class, 'cancel'])
        ->name('parallelruns.cancel');
    Route::get('admin/projects/{idProject}/parallel-runs/{parallelRun}/results', [ParallelRunScheduleController::class, 'results'])
        ->name('parallelruns.results');
    /* artifacts */
    Route::get(
        'admin/projects/{idProject}/performed-test-cycles/{performedTestCycleId}/artifacts',
        [ArtifactDescriptorController::class, 'index']
    )->name('artifacts.index');
    Route::get(
        'admin/projects/{idProject}/performed-test-cycles/{performedTestCycleId}/artifacts/{artifactDescriptor}',
        [ArtifactDescriptorController::class, 'show']
    )->name('artifacts.show');
    Route::get(
        'admin/projects/{idProject}/performed-test-cycles/{performedTestCycleId}/artifacts/{artifactDescriptor}/impact',
        [ArtifactDescriptorController::class, 'impact']
    )->name('artifacts.impact');
    Route::put(
        'admin/projects/{idProject}/performed-test-cycles/{performedTestCycleId}/artifacts/{artifactDescriptor}/legal-hold',
        [ArtifactDescriptorController::class, 'legalHold']
    )->name('artifacts.legalHold');
    Route::post(
        'admin/projects/{idProject}/performed-test-cycles/{performedTestCycleId}/artifacts/{artifactDescriptor}/delete-marker',
        [ArtifactDescriptorController::class, 'markDeleted']
    )->name('artifacts.markDeleted');
    Route::post(
        'admin/projects/{idProject}/performed-test-cycles/{performedTestCycleId}/artifacts/{artifactDescriptor}/archive',
        [ArtifactDescriptorController::class, 'archive']
    )->name('artifacts.archive');
    Route::post(
        'admin/projects/{idProject}/performed-test-cycles/{performedTestCycleId}/artifacts/{artifactDescriptor}/restore',
        [ArtifactDescriptorController::class, 'restore']
    )->name('artifacts.restore');
    /* testlauncher */
    Route::post('admin/launchtest', [TestLauncherController::class, 'launchTest'])
        ->name('testlauncher.launchTest');
    /* testcycles */
    Route::get('admin/testcycles/{idProject}', [TestCycleController::class, 'index'])
        ->name('testcycles.index');
    Route::get('admin/testcycles/{idProject}/{testcycle}', [TestCycleController::class, 'show'])
        ->name('testcycles.show');
    Route::put('admin/testcycles/{idProject}/{testcycle}', [TestCycleController::class, 'update'])
        ->name('testcycles.update');
    Route::post('admin/testcycles', [TestCycleController::class, 'store'])
        ->name('testcycles.store');
    /* importtest */
    Route::post('admin/importtest', [ImportSeleniumController::class, 'store'])
        ->name('importselenium.store');
    /* tests */
    Route::get('admin/tests/{idProject}', [TestController::class, 'index'])
        ->name('tests.index');
    Route::get('admin/tests/{idProject}/{test}', [TestController::class, 'show'])
        ->name('tests.show');
    Route::put('admin/tests/{idProject}/{test}', [TestController::class, 'update'])
        ->name('tests.update');
    Route::post('admin/tests', [TestController::class, 'store'])
        ->name('tests.store');
    /* steps */
    Route::post('admin/steps', [StepController::class, 'store'])
        ->name('steps.store');
    Route::get('admin/steps/{idProject}', [StepController::class, 'index'])
        ->name('steps.index');
    Route::get('admin/steps/{idProject}/{step}', [StepController::class, 'show'])
        ->name('steps.show');
    Route::put('admin/steps/{idProject}/{step}', [StepController::class, 'update'])
        ->name('steps.update');
    Route::delete('admin/steps/{idProject}/{environment}', [StepController::class, 'destroy'])
        ->name('steps.destroy');
    Route::post('admin/steps/{idProject}/updateorder', [StepController::class, 'updateorder'])
        ->name('steps.updateorder');
    /* plugins */
    Route::get('admin/plugins/{idProject}/{plugin}', [PluginController::class, 'show'])
        ->name('plugins.show');
    Route::get('admin/plugins/{idProject}', [PluginController::class, 'index'])
        ->name('plugins.index');
    Route::delete('admin/plugins/{idProject}/{plugin}', [PluginController::class, 'destroy'])
        ->name('plugins.destroy');
    Route::post('admin/plugins', [PluginController::class, 'store'])
        ->name('plugins.store');
    Route::put('admin/plugins/{idProject}/{step}', [PluginController::class, 'update'])
        ->name('plugins.update');
    /* environments */
    Route::get('admin/environments/{idProject}', [EnvironmentController::class, 'index'])
        ->name('environments.index');
    Route::get('admin/environments/{idProject}/{environment}', [EnvironmentController::class, 'show'])
        ->name('environments.show');
    Route::delete('admin/environments/{idProject}/{environment}', [EnvironmentController::class, 'destroy'])
        ->name('environments.destroy');
    Route::put('admin/environments/{idProject}/{environment}', [EnvironmentController::class, 'update'])
        ->name('environments.update');
    Route::post('admin/environments', [EnvironmentController::class, 'store'])
        ->name('environments.store');
    /* performed testcycles */
    Route::get('admin/testcyclesperfomed/{idTestCyclePerformed}', [PerformedTestCycleController::class, 'index'])
        ->name('testcyclesperfomed.index');
    /* performed test */
    Route::get('admin/testsperfomed/{idTestPerformed}', [PerformedTestController::class, 'index'])
        ->name('testsperfomed.index');
    /* performed step */
    Route::get('admin/stepsperfomed/{idTestPerformed}', [PerformedStepController::class, 'index'])
        ->name('testsperfomed.index');
    /* platforms */
    Route::get('admin/platforms/manageplatforms/{type}', [PlatformController::class, 'index'])
        ->name('platform.index');
    Route::post('admin/platforms/manageplatforms', [PlatformController::class, 'store'])
        ->name('platform.store');
    Route::put('admin/platforms/manageplatforms', [PlatformController::class, 'update'])
        ->name('platform.update');
    Route::delete('admin/platforms/manageplatforms/{type}/{id}', [PlatformController::class, 'delete'])
        ->name('platform.delete');
    /* platforms-status */
    Route::get('admin/platforms/status', [StatusController::class, 'index'])
        ->name('status.index');
    /* platforms-types */
    Route::get('admin/platforms/types', [TypeController::class, 'index'])
        ->name('types.index');
    /* platforms-os */
    Route::get('admin/platforms/os/{idType}', [OsController::class, 'index'])
        ->name('os.index');
    Route::post('admin/platforms/os', [OsController::class, 'store'])
        ->name('os.store');
    Route::put('admin/platforms/os', [OsController::class, 'update'])
        ->name('os.update');
    /* platforms-osversion */
    Route::get('admin/platforms/osversion/{idOs}', [VersionOsController::class, 'index'])
        ->name('osversion.index');
    Route::post('admin/platforms/osversion', [VersionOsController::class, 'store'])
        ->name('osversion.store');
    Route::put('admin/platforms/osversion', [VersionOsController::class, 'update'])
        ->name('osversion.update');
    /* platforms-browsers */
    Route::get('admin/platforms/browsers/{idOs}', [BrowserController::class, 'index'])
        ->name('browser.index');
    Route::post('admin/platforms/browsers', [BrowserController::class, 'store'])
        ->name('browser.store');
    Route::put('admin/platforms/browsers', [BrowserController::class, 'update'])
        ->name('browser.update');
    /* platforms-browserversions */
    Route::get('admin/platforms/browserversions/{idBrowser}', [VersionBrowserController::class, 'index'])
        ->name('versionbrowser.index');
    Route::post('admin/platforms/browserversions', [VersionBrowserController::class, 'store'])
        ->name('versionbrowser.store');
    Route::put('admin/platforms/browserversions', [VersionBrowserController::class, 'update'])
        ->name('versionbrowser.update');
    /* platforms-brands */
    Route::get('admin/platforms/brands', [BrandDeviceController::class, 'index'])
        ->name('brandevice.index');
    Route::post('admin/platforms/brands', [BrandDeviceController::class, 'store'])
        ->name('brandevice.store');
    Route::put('admin/platforms/brands', [BrandDeviceController::class, 'update'])
        ->name('brandevice.update');
    /* platforms-models */
    Route::get('admin/platforms/models/{idBrand}', [ModelDeviceController::class, 'index'])
        ->name('model.index');
    Route::post('admin/platforms/models', [ModelDeviceController::class, 'store'])
        ->name('model.store');
    Route::put('admin/platforms/models', [ModelDeviceController::class, 'update'])
        ->name('model.update');
    /* platforms-locations */
    Route::get('admin/platforms/locations', [LocationController::class, 'index'])
        ->name('location.index');
    Route::post('admin/platforms/locations', [LocationController::class, 'store'])
        ->name('location.store');
    Route::put('admin/platforms/locations', [LocationController::class, 'update'])
        ->name('location.update');
    Route::get('/user', function (Request $request) {
        return response()->json([
            'id' => $request->user()->id,
            'name' => $request->user()->name,
            'email' => $request->user()->email,
            'role' => $request->user()->role,
        ]);
    });
});

/* command line api */
Route::middleware('idelium.key')->prefix('ideliumcl')->group(function () {
    Route::post('agents/register', [AgentRegistrationController::class, 'register'])
        ->name('cl.agents.register');
    Route::get('projects/{idProject}/parallel-runs', [ParallelRunScheduleController::class, 'index'])
        ->name('cl.parallelruns.index');
    Route::post('projects/{idProject}/parallel-runs/matrix', [ParallelRunScheduleController::class, 'storeMatrix'])
        ->name('cl.parallelruns.storeMatrix');
    Route::post('projects/{idProject}/parallel-runs', [ParallelRunScheduleController::class, 'store'])
        ->name('cl.parallelruns.store');
    Route::get('projects/{idProject}/parallel-runs/{parallelRun}', [ParallelRunScheduleController::class, 'show'])
        ->name('cl.parallelruns.show');
    Route::post('projects/{idProject}/parallel-runs/{parallelRun}/claim', [ParallelRunScheduleController::class, 'claimWorker'])
        ->name('cl.parallelruns.claimWorker');
    Route::post('projects/{idProject}/parallel-runs/{parallelRun}/workers/{workerId}/heartbeat', [ParallelRunScheduleController::class, 'heartbeatWorker'])
        ->name('cl.parallelruns.heartbeatWorker');
    Route::post('projects/{idProject}/parallel-runs/{parallelRun}/tokens', [ParallelRunScheduleController::class, 'issueRunToken'])
        ->name('cl.parallelruns.issueRunToken');
    Route::post('projects/{idProject}/parallel-runs/{parallelRun}/tokens/{tokenId}/revoke', [ParallelRunScheduleController::class, 'revokeRunToken'])
        ->name('cl.parallelruns.revokeRunToken');
    Route::put('projects/{idProject}/parallel-runs/{parallelRun}/workers/{workerId}', [ParallelRunScheduleController::class, 'updateWorker'])
        ->name('cl.parallelruns.updateWorker');
    Route::post('projects/{idProject}/parallel-runs/{parallelRun}/cancel', [ParallelRunScheduleController::class, 'cancel'])
        ->name('cl.parallelruns.cancel');
    Route::get('projects/{idProject}/parallel-runs/{parallelRun}/results', [ParallelRunScheduleController::class, 'results'])
        ->name('cl.parallelruns.results');

    Route::get('testcycle/{idTestCycle}', [IdeliumClController::class, 'getTestCycle'])
        ->name('cl.getTestCycle');
    Route::get('test/{idTest}', [IdeliumClController::class, 'getTest'])
        ->name('cl.getTest');
    Route::get('step/{idStep}', [IdeliumClController::class, 'getStep'])
        ->name('cl.getStep');
    Route::get('plugins/{idProject}', [IdeliumClController::class, 'getPlugins'])
        ->name('cl.getPlugins');
    Route::get('plugin/{idPlugin}', [IdeliumClController::class, 'getPlugin'])
        ->name('cl.getPlugin');
    Route::get('environments/{idProject}', [IdeliumClController::class, 'getEnvironments'])
        ->name('cl.getEnvironments');
    Route::get('environment/{idEnvironment}', [IdeliumClController::class, 'getEnvironment'])
        ->name('cl.getEnvironment');

    Route::post('testcycle', [IdeliumInsertClController::class, 'createFolder'])
        ->name('cl.createFolder');
    Route::post('test', [IdeliumInsertClController::class, 'createTest'])
        ->name('cl.createTest');
    Route::put('test', [IdeliumInsertClController::class, 'updateTest'])
        ->name('cl.updateTest');
    Route::post('step', [IdeliumInsertClController::class, 'createStep'])
        ->name('cl.createStep');
    Route::put('step', [IdeliumInsertClController::class, 'updateStep'])
        ->name('cl.updateStep');
});